requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.8
python-chess>=1.3
PyYAML>=4.2b1
requests>=2.21.0
urllib3==1.24.2
backoff==1.7.1
beautifulsoup4